requirements.txt
atomicwrites==1.2.1
attrs==18.2.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
chardet==3.0.4
colorama==0.4.1
fpdf==1.7.2
idna==2.8
more-itertools==4.3.0
pdfkit==0.6.1
pdfminer3k==1.3.1
Pillow==5.3.0
pluggy==0.8.0
ply==3.11
py==1.7.0
pytest==4.0.2
python-pdf==0.36
reportlab==3.5.12
requests==2.21.0
selenium==3.141.0
six==1.12.0
urllib3==1.24.1
wkhtmltopdf==0.2
